Expose server
What good is a server if clients cannot access it?

This seems like something wrong with Docker's default
bridge network configuration; traffic should be bridged
from the host system to the container, but I'm not sure
about how to set it up. In the meantime, this is a suitable
workaround.